+++ packages/flutter_markdown/lib/src/widget.dart
@@ -545,6 +545,30 @@ class Markdown extends MarkdownWidget {
   }
 }
 
+/// Parse [task list items](https://github.github.com/gfm/#task-list-items-extension-).
+/// This class is no longer used as Markdown now supports checkbox syntax natively
+/// FIXME: Remove class entirely at some point in the future.
+@Deprecated(
+    'Use [OrderedListWithCheckBoxSyntax] or [UnorderedListWithCheckBoxSyntax]')
+class TaskListSyntax extends md.InlineSyntax {
+  /// Creates a new instance.
+  @Deprecated(
+      'Use [OrderedListWithCheckBoxSyntax] or [UnorderedListWithCheckBoxSyntax]')
+  TaskListSyntax() : super(_pattern);
+
+  static const String _pattern = r'^ *\[([ xX])\] +';
+
+  @override
+  bool onMatch(md.InlineParser parser, Match match) {
+    final md.Element el = md.Element.withTag('input');
+    el.attributes['type'] = 'checkbox';
+    el.attributes['disabled'] = 'true';
+    el.attributes['checked'] = '${match[1]!.trim().isNotEmpty}';
+    parser.addNode(el);
+    return true;
+  }
+}
+
 /// An interface for an padding builder for element.
 abstract class MarkdownPaddingBuilder {
   /// Called when an Element has been reached, before its children have been